refactor: 사용자입력 클래스 합치기
