+++ src/view/InputView.js
@@ -1,41 +1,57 @@
 import { Console } from '@woowacourse/mission-utils';
 import { ERROR_MESSAGE, INPUT_MESSAGE } from '../constants/message.js';
-import { LOTTO } from '../constants/lotto.js';
-import { COMMON } from '../constants/common.js';
 import { validation } from '../utils/validation.js';
 import ValidateError from '../error/ValidateError.js';
 
-class Lottery {
-  #winningNumbers;
-  #bonusNumber;
-
-  async getLottoWinningNumbers() {
+class InputView {
+  async getLottoPurchasePrice() {
     while (true) {
       try {
-        Console.print(INPUT_MESSAGE.lottoWinningNumbersDescription);
-        const lottoWinningNumbers = await Console.readLineAsync(INPUT_MESSAGE.lottoWnningNumbers);
+        const lottoPurchasePrice = await Console.readLineAsync(INPUT_MESSAGE.lottoPurchasePrice);
 
-        this.#validateNumbers(lottoWinningNumbers);
+        this.#validate(lottoPurchasePrice);
 
-        this.#winningNumbers = lottoWinningNumbers;
-
-        break;
+        return Number(lottoPurchasePrice);
       } catch (error) {
         Console.print(error);
       }
     }
   }
 
-  async getLottoBonusNumber() {
+  #validate(price) {
+    if (validation.isEmpty(price)) {
+      throw new ValidateError(ERROR_MESSAGE.empty);
+    }
+
+    const numberPrice = Number(price);
+
+    this.#numberValidate(numberPrice);
+  }
+
+  #numberValidate(price) {
+    if (validation.isNumberZero(price)) {
+      throw new ValidateError(ERROR_MESSAGE.notZero);
+    }
+
+    if (isNaN(price)) {
+      throw new ValidateError(ERROR_MESSAGE.notNumber);
+    }
+
+    if (validation.isNotLottoPurchagePriceRange(price)) {
+      throw new ValidateError(ERROR_MESSAGE.lottoPurchasePriceRange);
+    }
+  }
+
+  async getLottoWinningNumbers() {
     while (true) {
       try {
-        const lottoBonusNumber = await Console.readLineAsync(INPUT_MESSAGE.lottoBonusNumber);
+        Console.print(INPUT_MESSAGE.lottoWinningNumbersDescription);
 
-        this.#validateNumber(Number(lottoBonusNumber));
+        const lottoWinningNumbers = await Console.readLineAsync(INPUT_MESSAGE.lottoWnningNumbers);
 
-        this.#bonusNumber = lottoBonusNumber;
+        this.#validateNumbers(lottoWinningNumbers);
 
-        break;
+        return lottoWinningNumbers;
       } catch (error) {
         Console.print(error);
       }
@@ -72,6 +88,20 @@ class Lottery {
     }
   }
 
+  async getLottoBonusNumber() {
+    while (true) {
+      try {
+        const lottoBonusNumber = await Console.readLineAsync(INPUT_MESSAGE.lottoBonusNumber);
+
+        this.#validateNumber(Number(lottoBonusNumber));
+
+        return lottoBonusNumber;
+      } catch (error) {
+        Console.print(error);
+      }
+    }
+  }
+
   #validateNumber(number) {
     if (validation.isEmpty(number)) {
       throw new ValidateError(ERROR_MESSAGE.empty);
@@ -81,19 +111,12 @@ class Lottery {
       throw new ValidateError(ERROR_MESSAGE.lottoNumberRangee);
     }
 
-    const isWinningNumbersIncludeBonusNumber = this.#winningNumbers.includes(number);
+    // const isWinningNumbersIncludeBonusNumber = this.#winningNumbers.includes(number);
 
     if (isWinningNumbersIncludeBonusNumber) {
       throw new ValidateError(ERROR_MESSAGE.bonusNumberIncludeWinningNumber);
     }
   }
-
-  getWinningNumbers() {
-    return {
-      winningNumbers: this.#winningNumbers,
-      bonusNumber: this.#bonusNumber,
-    };
-  }
 }
 
-export default Lottery;
+export default InputView;
